Move definitions to separated projects
