+++ Reflection/CollectionActivation.cs
@@ -27,7 +27,7 @@ namespace YggdrAshill.Ragnarok
                 // TODO: Type.IsInstanceOfType(object)?
                 if (!elementType.IsAssignableFrom(parameterType))
                 {
-                    throw new RagnarokArgumentException(parameterType, $"{parameterType} is not assignable from {elementType}.");
+                    throw new RagnarokReflectionException(parameterType, $"{parameterType} is not assignable from {elementType}.");
                 }
 
                 array.SetValue(parameter, index);
