+++ Reflection/ReflectionFieldInfusion.cs
@@ -23,11 +23,11 @@ namespace YggdrAshill.Ragnarok
 
             if (!implementedType.IsInstanceOfType(instance))
             {
-                throw new RagnarokArgumentException(implementedType, $"{instance} is not {implementedType}.");
+                throw new RagnarokReflectionException(implementedType, $"{instance} is not {implementedType}.");
             }
             if (fieldList.Length != parameterList.Length)
             {
-                throw new RagnarokArgumentException(implementedType, nameof(parameterList));
+                throw new RagnarokReflectionException(implementedType, nameof(parameterList));
             }
 
             for (var index = 0; index < fieldList.Length; index++)
@@ -41,7 +41,7 @@ namespace YggdrAshill.Ragnarok
                 // TODO: Type.IsInstanceOfType(object)?
                 if (!fieldType.IsAssignableFrom(parameterType))
                 {
-                    throw new RagnarokArgumentException(parameterType, $"{parameterType} is not assignable from {fieldType}.");
+                    throw new RagnarokReflectionException(parameterType, $"{parameterType} is not assignable from {fieldType}.");
                 }
 
                 field.SetValue(instance, parameter);
