+++ Reflection/ReflectionSolver.cs
