+++ YggdrAshill.Ragnarok/YggdrAshill.Ragnarok.csproj
@@ -8,10 +8,12 @@
 
   <ItemGroup>
     <ProjectReference Include="..\Construction\Construction.csproj" />
+    <ProjectReference Include="..\Expression\Expression.csproj" />
     <ProjectReference Include="..\Fabrication\Fabrication.csproj" />
     <ProjectReference Include="..\Hierarchization\Hierarchization.csproj" />
     <ProjectReference Include="..\Materialization\Materialization.csproj" />
     <ProjectReference Include="..\Motorization\Motorization.csproj" />
+    <ProjectReference Include="..\Reflection\Reflection.csproj" />
   </ItemGroup>
 
 </Project>
